@@ -1,14 +1,13 @@
 import 'package:google_speech/config/recognition_config.dart';
 import 'package:google_speech/generated/google/cloud/speech/v1/cloud_speech.pb.dart'
     as _cs;
-import 'package:meta/meta.dart';
 
 class RecognitionConfig {
   /// Provides information to the recognizer
   /// that specifies how to process the request.
   RecognitionConfig({
-    @required this.encoding,
-    @required this.languageCode,
+    required this.encoding,
+    required this.languageCode,
     this.sampleRateHertz,
     this.audioChannelCount = 1,
     this.enableSeparateRecognitionPerChannel = false,
@@ -38,7 +37,7 @@ class RecognitionConfig {
   _cs.RecognitionConfig toConfig() => (_cs.RecognitionConfig()
     ..encoding = _encoding(encoding)
     ..languageCode = languageCode
-    ..sampleRateHertz = sampleRateHertz
+    ..sampleRateHertz = sampleRateHertz!
     ..audioChannelCount = audioChannelCount
     ..enableSeparateRecognitionPerChannel = enableSeparateRecognitionPerChannel
     ..maxAlternatives = maxAlternatives
@@ -100,7 +99,7 @@ class RecognitionConfig {
   /// (instead of re-sampling). This field is optional for FLAC and WAV audio
   /// files, but is required for all other audio formats.
   /// For details, see [AudioEncoding].
-  int sampleRateHertz;
+  int? sampleRateHertz;
 
   /// The number of channels in the input audio data. ONLY set this for
   /// MULTI-CHANNEL recognition. Valid values for LINEAR16 and FLAC are 1-8.
@@ -183,10 +182,10 @@ class RecognitionConfig {
   /// identify the speakers in the conversation over time.
   /// For non-streaming requests, the diarization results will be provided only
   /// in the top alternative of the FINAL SpeechRecognitionResult.
-  final _cs.SpeakerDiarizationConfig diarizationConfig;
+  final _cs.SpeakerDiarizationConfig? diarizationConfig;
 
   /// Metadata regarding this request.
-  final _cs.RecognitionMetadata recognitionMetadata;
+  final _cs.RecognitionMetadata? recognitionMetadata;
 }
 
 class SpeechContext {
