@@ -5,7 +5,7 @@ abstract class UnsupportedFileException implements IOException {
   final String message;
 
   /// File that was transferred during initialization.
-  final File file;
+  final File? file;
 
   /// Will be thrown if a [File] is passed when initializing a [ServiceAccount]
   /// which does not contain the necessary content to
@@ -19,10 +19,10 @@ abstract class UnsupportedFileException implements IOException {
     if (message.isNotEmpty) {
       sb.write(': $message');
       if (file?.path != null) {
-        sb.write(", path = '${file.path}'");
+        sb.write(", path = '${file?.path}'");
       }
     } else if (file?.path != null) {
-      sb.write(': ${file.path}');
+      sb.write(': ${file?.path}');
     }
     return sb.toString();
   }
@@ -31,7 +31,7 @@ abstract class UnsupportedFileException implements IOException {
 class FileNotFoundException extends UnsupportedFileException {
   /// Will be thrown, if a [File] is passed, which is not found on
   /// the file system.
-  FileNotFoundException(File file)
+  FileNotFoundException(File? file)
       : super(
             'The transferred file could not be found. '
             'Make sure that the file ${file?.path} exists.',
@@ -40,7 +40,7 @@ class FileNotFoundException extends UnsupportedFileException {
 
 class UnsupportedFileExtensionException extends UnsupportedFileException {
   /// Will be thrown, if a [File] is passed, which has no json extension.
-  UnsupportedFileExtensionException(File file)
+  UnsupportedFileExtensionException(File? file)
       : super(
             'The file extension ${file?.path?.split('.')?.last} is not '
             'supported. Make sure that the transferred file is a valid Json '
