@@ -2,7 +2,7 @@
 //  Generated code. Do not modify.
 //  source: google/protobuf/timestamp.proto
 //
-// @dart = 2.3
+
 // ignore_for_file: camel_case_types,non_constant_identifier_names,library_prefixes,unused_import,unused_shown_name,return_of_invalid_type
 
 import 'dart:core' as $core;
@@ -32,7 +32,7 @@ class Timestamp extends $pb.GeneratedMessage with $mixin.TimestampMixin {
       create()..mergeFromJson(i, r);
   Timestamp clone() => Timestamp()..mergeFromMessage(this);
   Timestamp copyWith(void Function(Timestamp) updates) =>
-      super.copyWith((message) => updates(message as Timestamp));
+      super.copyWith((message) => updates(message as Timestamp)) as Timestamp;
   $pb.BuilderInfo get info_ => _i;
   @$core.pragma('dart2js:noInline')
   static Timestamp create() => Timestamp._();
@@ -41,7 +41,7 @@ class Timestamp extends $pb.GeneratedMessage with $mixin.TimestampMixin {
   @$core.pragma('dart2js:noInline')
   static Timestamp getDefault() =>
       _defaultInstance ??= $pb.GeneratedMessage.$_defaultFor<Timestamp>(create);
-  static Timestamp _defaultInstance;
+  static Timestamp? _defaultInstance;
 
   @$pb.TagNumber(1)
   $fixnum.Int64 get seconds => $_getI64(0);
