@@ -2,7 +2,7 @@
 //  Generated code. Do not modify.
 //  source: google/rpc/status.proto
 //
-// @dart = 2.3
+
 // ignore_for_file: camel_case_types,non_constant_identifier_names,library_prefixes,unused_import,unused_shown_name,return_of_invalid_type
 
 import 'dart:core' as $core;
@@ -29,7 +29,7 @@ class Status extends $pb.GeneratedMessage {
       create()..mergeFromJson(i, r);
   Status clone() => Status()..mergeFromMessage(this);
   Status copyWith(void Function(Status) updates) =>
-      super.copyWith((message) => updates(message as Status));
+      super.copyWith((message) => updates(message as Status)) as Status;
   $pb.BuilderInfo get info_ => _i;
   @$core.pragma('dart2js:noInline')
   static Status create() => Status._();
@@ -38,7 +38,7 @@ class Status extends $pb.GeneratedMessage {
   @$core.pragma('dart2js:noInline')
   static Status getDefault() =>
       _defaultInstance ??= $pb.GeneratedMessage.$_defaultFor<Status>(create);
-  static Status _defaultInstance;
+  static Status? _defaultInstance;
 
   @$pb.TagNumber(1)
   $core.int get code => $_getIZ(0);
