@@ -2,7 +2,7 @@
 //  Generated code. Do not modify.
 //  source: google/longrunning/operations.proto
 //
-// @dart = 2.3
+
 // ignore_for_file: camel_case_types,non_constant_identifier_names,library_prefixes,unused_import,unused_shown_name,return_of_invalid_type
 
 import 'dart:async' as $async;
@@ -42,12 +42,12 @@ class OperationsClient extends $grpc.Client {
           ($0.WaitOperationRequest value) => value.writeToBuffer(),
           ($core.List<$core.int> value) => $0.Operation.fromBuffer(value));
 
-  OperationsClient($grpc.ClientChannel channel, {$grpc.CallOptions options})
+  OperationsClient($grpc.ClientChannel channel, {$grpc.CallOptions? options})
       : super(channel, options: options);
 
   $grpc.ResponseFuture<$0.ListOperationsResponse> listOperations(
       $0.ListOperationsRequest request,
-      {$grpc.CallOptions options}) {
+      {$grpc.CallOptions? options}) {
     final call = $createCall(
         _$listOperations, $async.Stream.fromIterable([request]),
         options: options);
@@ -56,7 +56,7 @@ class OperationsClient extends $grpc.Client {
 
   $grpc.ResponseFuture<$0.Operation> getOperation(
       $0.GetOperationRequest request,
-      {$grpc.CallOptions options}) {
+      {$grpc.CallOptions? options}) {
     final call = $createCall(
         _$getOperation, $async.Stream.fromIterable([request]),
         options: options);
@@ -65,7 +65,7 @@ class OperationsClient extends $grpc.Client {
 
   $grpc.ResponseFuture<$1.Empty> deleteOperation(
       $0.DeleteOperationRequest request,
-      {$grpc.CallOptions options}) {
+      {$grpc.CallOptions? options}) {
     final call = $createCall(
         _$deleteOperation, $async.Stream.fromIterable([request]),
         options: options);
@@ -74,7 +74,7 @@ class OperationsClient extends $grpc.Client {
 
   $grpc.ResponseFuture<$1.Empty> cancelOperation(
       $0.CancelOperationRequest request,
-      {$grpc.CallOptions options}) {
+      {$grpc.CallOptions? options}) {
     final call = $createCall(
         _$cancelOperation, $async.Stream.fromIterable([request]),
         options: options);
@@ -83,7 +83,7 @@ class OperationsClient extends $grpc.Client {
 
   $grpc.ResponseFuture<$0.Operation> waitOperation(
       $0.WaitOperationRequest request,
-      {$grpc.CallOptions options}) {
+      {$grpc.CallOptions? options}) {
     final call = $createCall(
         _$waitOperation, $async.Stream.fromIterable([request]),
         options: options);
